model_1/v2_vertex/requirements.txt
google-cloud-aiplatform>=1.54.0
vertexai>=1.54.0
google-auth
google-auth-oauthlib
google-auth-httplib2

langchain-core>=0.1.40

Pillow
pdf2image

pydantic>=2.0
python-multipart
python-Levenshtein
